Migrate more tests to use pdata
